gst/va: move HEVC encoder test classes to lib
